Facelift
This commit is a new look for the {code}bridge site, that can be a
starting point to build upon. I updated to Bootstrap 3.3.5, switched to
Google hosted libraries for jQuery (better caching, increased
parallelism and decreased latency). Then updated Jekyll parameters to
latest version (including _layout folder). Added Open Graph meta tags
for better sharing, which can be adjusted in the config file. You will
also find the resources I used for the facelift in the config file. The
facelift itself was minor changes to the content, more focussed on
making it more responsive in a full page layout. It is a starting point
and requires fine tuning. The original site is 7zipped in the backup
folder, which I did before deleting older content.